HTTP/3: Downgraded error log to report QUIC stream creation failure while closing session
The HTTP/3 module closes a session on idle timeout according to RFC 9000 section 10.1.
The remote peer may also close the session at the same time. In that case, closing session on
the local peer may encounter errors in creating streams. The errors in this condition, however,
do not harm actually. The fix downgrades the log level to info in such a case.